requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.21
django-bootstrap5==25.1
django-extensions==4.1
djangorestframework==3.16.0
dotenv==0.9.9
exceptiongroup==1.3.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
pytest-django==4.5.2
python-dotenv==1.1.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
